+++ src/run/constants.py
@@ -164,6 +164,6 @@ SEVEN_ZIP_PATH = read_ini_config(ini_file_path, "unzip_config", "SEVEN_ZIP_PATH"
 # unzip zip file password
 PASSWORD = read_ini_config(ini_file_path, "unzip_config", "PASSWORD")
 
-sis_server_version = "v1.1.4-beta.2.240428"
-build_date = "2024-04-28 18:00"
-publish_date = "2024-04-28 18:30"
+sis_server_version = "v1.1.4.240430"
+build_date = "2024-04-30 18:00"
+publish_date = "2024-04-30 18:30"